V2024.12.8

+++ Mips/src/structure/data/ByteStructure.cpp
@@ -16,12 +16,12 @@ string ByteStructure::toMips() {
         //没有初始值
         return label + ": .byte " + value;
     }
-    int offset = stoi(value) - static_cast<int>(initValues.size());
+    int offset = stoll(value) - static_cast<int>(initValues.size());
     ostringstream oss;
     oss<<label + ": .byte ";
     for (int i = 0; i < initValues.size(); i++) {
         oss << initValues[i];
-        if(i+1<stoi(value)) {
+        if(i+1<stoll(value)) {
             oss<<", ";
         }
     }
